split tokenization utils to isolate heavy deps
